+++ mmyolo/models/backbones/__init__.py
@@ -1,13 +1,12 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .base_backbone import BaseBackbone
 from .csp_darknet import YOLOv5CSPDarknet, YOLOXCSPDarknet
-from .csp_resnet import CSPResNet, PPYOLOEBasicBlock
+from .csp_resnet import PPYOLOECSPResNet
 from .cspnext import CSPNeXt
 from .efficient_rep import YOLOv6EfficientRep
 from .yolov7_backbone import YOLOv7Backbone
 
 __all__ = [
     'YOLOv5CSPDarknet', 'BaseBackbone', 'YOLOv6EfficientRep',
-    'YOLOXCSPDarknet', 'CSPNeXt', 'YOLOv7Backbone', 'CSPResNet',
-    'PPYOLOEBasicBlock'
+    'YOLOXCSPDarknet', 'CSPNeXt', 'YOLOv7Backbone', 'PPYOLOECSPResNet'
 ]
