+++ mmyolo/models/layers/__init__.py
@@ -1,9 +1,12 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .ema import ExpMomentumEMA
-from .yolo_bricks import (ELANBlock, MaxPoolAndStrideConvBlock, RepStageBlock,
-                          RepVGGBlock, SPPFBottleneck, SPPFCSPBlock)
+from .yolo_bricks import (EffectiveSELayer, ELANBlock,
+                          MaxPoolAndStrideConvBlock, PPYOLOEBasicBlock,
+                          RepStageBlock, RepVGGBlock, SPPFBottleneck,
+                          SPPFCSPBlock)
 
 __all__ = [
     'SPPFBottleneck', 'RepVGGBlock', 'RepStageBlock', 'ExpMomentumEMA',
-    'ELANBlock', 'MaxPoolAndStrideConvBlock', 'SPPFCSPBlock'
+    'ELANBlock', 'MaxPoolAndStrideConvBlock', 'SPPFCSPBlock',
+    'PPYOLOEBasicBlock', 'EffectiveSELayer'
 ]
