+++ mmyolo/models/layers/yolo_bricks.py
@@ -365,6 +365,7 @@ class RepStageBlock(nn.Module):
         return x
 
 
+@MODELS.register_module()
 class EffectiveSELayer(nn.Module):
     """Effective Squeeze-Excitation.
 
@@ -779,7 +780,7 @@ class SPP(nn.Module):
     """
 
     def __init__(self,
-                 input_channels: int,
+                 in_channels: int,
                  out_channels: int,
                  kernel_size: int,
                  pool_kernel_sizes: Tuple[int] = (5, 9, 13),
@@ -793,7 +794,7 @@ class SPP(nn.Module):
             for ks in pool_kernel_sizes
         ])
         self.conv = ConvModule(
-            input_channels,
+            in_channels,
             out_channels,
             kernel_size,
             padding=kernel_size // 2,
@@ -812,4 +813,203 @@ class SPP(nn.Module):
         y = torch.cat(outs, axis=1)
 
         y = self.conv(y)
+        return y
+
+
+@MODELS.register_module()
+class PPYOLOEBasicBlock(nn.Module):
+    """PPYOLOE Backbone BasicBlock.
+
+    Args:
+         in_channels (int): The input channels of this Module.
+         out_channels (int): The output channels of this Module.
+         norm_cfg (dict): Config dict for normalization layer.
+             Defaults to dict(type='BN', momentum=0.1, eps=1e-5).
+         act_cfg (dict): Config dict for activation layer.
+             Defaults to dict(type='SiLU', inplace=True).
+         shortcut (bool): Whether to add inputs and outputs together
+         at the end of this layer. Defaults to True.
+         use_alpha (bool): Whether to use `alpha` parameter at 1x1 conv.
+    """
+
+    def __init__(self,
+                 in_channels: int,
+                 out_channels: int,
+                 norm_cfg: ConfigType = dict(
+                     type='BN', momentum=0.1, eps=1e-5),
+                 act_cfg: ConfigType = dict(type='SiLU', inplace=True),
+                 shortcut: bool = True,
+                 use_alpha: bool = False):
+        super().__init__()
+        assert act_cfg is None or isinstance(act_cfg, dict)
+        self.conv1 = ConvModule(
+            in_channels,
+            out_channels,
+            3,
+            stride=1,
+            padding=1,
+            norm_cfg=norm_cfg,
+            act_cfg=act_cfg)
+
+        self.conv2 = RepVGGBlock(
+            out_channels,
+            out_channels,
+            use_alpha=use_alpha,
+            act_cfg=act_cfg,
+            norm_cfg=norm_cfg,
+            use_bn_first=False)
+        self.shortcut = shortcut
+
+    def forward(self, x: Tensor) -> Tensor:
+        """Forward process.
+        Args:
+            inputs (Tensor): The input tensor.
+
+        Returns:
+            Tensor: The output tensor.
+        """
+        y = self.conv1(x)
+        y = self.conv2(y)
+        if self.shortcut:
+            return x + y
+        else:
+            return y
+
+
+class CSPResLayer(nn.Module):
+    """PPYOLOE Backbone Stage.
+
+    Args:
+        in_channels (int): The input channels of this Module.
+        out_channels (int): The output channels of this Module.
+        num_block (int): Number of blocks in this stage.
+        block_cfg (dict): Config dict for block. Default config is
+            suitable for PPYOLOE+ backbone. And in PPYOLOE neck,
+            block_cfg is set to dict(type='PPYOLOEBasicBlock',
+            shortcut=False, use_alpha=False). Defaults to
+            dict(type='PPYOLOEBasicBlock', shortcut=True, use_alpha=True).
+        stride (int): Stride of the convolution. In backbone, the stride
+            must be set to 2. In neck, the stride must be set to 1.
+            Defaults to 1.
+        norm_cfg (dict): Config dict for normalization layer.
+            Defaults to dict(type='BN', momentum=0.1, eps=1e-5).
+        act_cfg (dict): Config dict for activation layer.
+            Defaults to dict(type='SiLU', inplace=True).
+        effective_se_cfg (dict, optional): Config dict for `EffectiveSELayer`.
+            Defaults to dict(type='EffectiveSELayer',
+            act_cfg=dict(type='HSigmoid')).
+        use_spp (bool): Whether to use `SPP` layer.
+            Defaults to False.
+    """
+
+    def __init__(self,
+                 in_channels: int,
+                 out_channels: int,
+                 num_block: int,
+                 block_cfg: ConfigType = dict(
+                     type='PPYOLOEBasicBlock', shortcut=True, use_alpha=True),
+                 stride: int = 1,
+                 norm_cfg: ConfigType = dict(
+                     type='BN', momentum=0.1, eps=1e-5),
+                 act_cfg: ConfigType = dict(type='SiLU', inplace=True),
+                 effective_se_cfg: OptMultiConfig = dict(
+                     type='EffectiveSELayer', act_cfg=dict(type='HSigmoid')),
+                 use_spp: bool = False):
+        super().__init__()
+
+        self.num_block = num_block
+        self.block_cfg = block_cfg
+        self.norm_cfg = norm_cfg
+        self.act_cfg = act_cfg
+        self.use_spp = use_spp
+        assert effective_se_cfg is None or isinstance(effective_se_cfg, dict)
+
+        if stride == 2:
+            conv1_in_channels = conv2_in_channels = conv3_in_channels = (
+                in_channels + out_channels) // 2
+            blocks_channels = conv1_in_channels // 2
+            self.conv_down = ConvModule(
+                in_channels,
+                conv1_in_channels,
+                3,
+                stride=2,
+                padding=1,
+                norm_cfg=norm_cfg,
+                act_cfg=act_cfg)
+        else:
+            conv1_in_channels = conv2_in_channels = in_channels
+            conv3_in_channels = out_channels
+            blocks_channels = out_channels // 2
+            self.conv_down = None
+
+        self.conv1 = ConvModule(
+            conv1_in_channels,
+            blocks_channels,
+            1,
+            norm_cfg=norm_cfg,
+            act_cfg=act_cfg)
+
+        self.conv2 = ConvModule(
+            conv2_in_channels,
+            blocks_channels,
+            1,
+            norm_cfg=norm_cfg,
+            act_cfg=act_cfg)
+
+        self.blocks = self.build_blocks_layer(blocks_channels)
+
+        self.conv3 = ConvModule(
+            conv3_in_channels,
+            out_channels,
+            1,
+            norm_cfg=norm_cfg,
+            act_cfg=act_cfg)
+
+        if effective_se_cfg:
+            effective_se_cfg = effective_se_cfg.copy()
+            effective_se_cfg['channels'] = blocks_channels * 2
+            self.attn = MODELS.build(effective_se_cfg)
+        else:
+            self.attn = None
+
+    def build_blocks_layer(self, blocks_channels: int) -> nn.Module:
+        """Build blocks layer.
+
+        Args:
+            blocks_channels: The channels of this Module.
+        """
+        blocks = nn.Sequential()
+        block_cfg = self.block_cfg.copy()
+        block_cfg.update(
+            dict(in_channels=blocks_channels, out_channels=blocks_channels))
+        block_cfg.setdefault('norm_cfg', self.norm_cfg)
+        block_cfg.setdefault('act_cfg', self.act_cfg)
+
+        for i in range(self.num_block):
+            blocks.add_module(str(i), MODELS.build(block_cfg))
+
+            if i == (self.num_block - 1) // 2 and self.use_spp:
+                blocks.add_module(
+                    'spp',
+                    SPP(blocks_channels * 4,
+                        blocks_channels,
+                        1, [5, 9, 13],
+                        norm_cfg=self.norm_cfg,
+                        act_cfg=self.act_cfg))
+
+        return blocks
+
+    def forward(self, x: Tensor) -> Tensor:
+        """Forward process
+         Args:
+             x (Tensor): The input tensor.
+         """
+        if self.conv_down is not None:
+            x = self.conv_down(x)
+        y1 = self.conv1(x)
+        y2 = self.blocks(self.conv2(x))
+        y = torch.cat([y1, y2], axis=1)
+        if self.attn is not None:
+            y = self.attn(y)
+        y = self.conv3(y)
         return y
