+++ mmyolo/models/necks/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .base_yolo_neck import BaseYOLONeck
 from .cspnext_pafpn import CSPNeXtPAFPN
-from .ppyoloe_csppan import PPYOLOECSPPAN
+from .ppyoloe_csppan import PPYOLOECSPPAFPN
 from .yolov5_pafpn import YOLOv5PAFPN
 from .yolov6_pafpn import YOLOv6RepPAFPN
 from .yolov7_pafpn import YOLOv7PAFPN
@@ -9,5 +9,5 @@ from .yolox_pafpn import YOLOXPAFPN
 
 __all__ = [
     'YOLOv5PAFPN', 'BaseYOLONeck', 'YOLOv6RepPAFPN', 'YOLOXPAFPN',
-    'CSPNeXtPAFPN', 'YOLOv7PAFPN', 'PPYOLOECSPPAN'
+    'CSPNeXtPAFPN', 'YOLOv7PAFPN', 'PPYOLOECSPPAFPN'
 ]
