+++ mmyolo/models/necks/ppyoloe_csppan.py
@@ -1,17 +1,17 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from typing import List, Union
+from typing import List
 
 import torch.nn as nn
 from mmcv.cnn import ConvModule
 from mmdet.utils import ConfigType, OptMultiConfig
 
-from mmyolo.models.backbones.csp_resnet import CSPResStage
+from mmyolo.models.backbones.csp_resnet import CSPResLayer
 from mmyolo.models.necks import BaseYOLONeck
 from mmyolo.registry import MODELS
 
 
 @MODELS.register_module()
-class PPYOLOECSPPAN(BaseYOLONeck):
+class PPYOLOECSPPAFPN(BaseYOLONeck):
     """CSPPAN in PPYOLOE.
 
     Args:
@@ -23,55 +23,51 @@ class PPYOLOECSPPAN(BaseYOLONeck):
         widen_factor (float): Width multiplier, multiply number of
             channels in each layer by this amount. Defaults to 1.0.
         freeze_all(bool): Whether to freeze the model.
+        num_csplayer (int): Number of `CSPResLayer` in per layer.
+            Defaults to 1.
+        num_blocks_per_layer (int): Number of blocks per `CSPResLayer`.
+            Defaults to 3.
         block_cfg (dict): Config dict for block. Defaults to
             dict(type='PPYOLOEBasicBlock', shortcut=True, use_alpha=False)
         norm_cfg (dict): Config dict for normalization layer.
             Defaults to dict(type='BN', momentum=0.1, eps=1e-5).
         act_cfg (dict): Config dict for activation layer.
             Defaults to dict(type='SiLU', inplace=True).
+        drop_block_cfg (dict, optional): Drop block config.
+            Defaults to None. If you want to use Drop block after
+            `CSPResLayer`, you can set this para as
+            dict(type='mmdet.DropBlock', drop_prob=0.1,
+            block_size=3, warm_iters=0).
         init_cfg (dict or list[dict], optional): Initialization config dict.
             Defaults to None.
-        num_stages_per_layer (int): Number of stages per layer.
-            Defaults to 1.
-        num_blocks_per_stage (int): Number of blocks per stage.
-            Defaults to 3.
         use_spp (bool): Whether to use `SPP` in reduce layer.
             Defaults to False.
-        use_drop_block (bool): Whether to use `DropBlock` after `CSPStage`.
-            Defaults to False.
-        drop_block_cfg (dict): If use_drop_block is True, build drop_block
-            by this cfg. Defaults to dict(type='mmdet.DropBlock',drop_prob=0.1,
-            block_size=3, warm_iters=0)
     """
 
     def __init__(self,
                  in_channels: List[int] = [256, 512, 1024],
-                 out_channels: Union[int, List[int]] = [256, 512, 1024],
+                 out_channels: List[int] = [256, 512, 1024],
                  deepen_factor: float = 1.0,
                  widen_factor: float = 1.0,
                  freeze_all: bool = False,
+                 num_csplayer: int = 1,
+                 num_blocks_per_layer: int = 3,
                  block_cfg: ConfigType = dict(
                      type='PPYOLOEBasicBlock', shortcut=False,
                      use_alpha=False),
                  norm_cfg: ConfigType = dict(
                      type='BN', momentum=0.1, eps=1e-5),
                  act_cfg: ConfigType = dict(type='SiLU', inplace=True),
+                 drop_block_cfg: ConfigType = None,
                  init_cfg: OptMultiConfig = None,
-                 num_stages_per_layer: int = 1,
-                 num_blocks_per_stage: int = 3,
-                 use_spp: bool = False,
-                 use_drop_block: bool = False,
-                 drop_block_cfg: ConfigType = dict(
-                     type='mmdet.DropBlock',
-                     drop_prob=0.1,
-                     block_size=3,
-                     warm_iters=0)):
+                 use_spp: bool = False):
         self.block_cfg = block_cfg
-        self.num_stages_per_layer = num_stages_per_layer
-        self.num_blocks_per_stage = round(num_blocks_per_stage * deepen_factor)
+        self.num_csplayer = num_csplayer
+        self.num_blocks_per_layer = round(num_blocks_per_layer * deepen_factor)
+        # Only use spp in last reduce_layer, if use_spp=True.
         self.use_spp = use_spp
-        self.use_drop_block = use_drop_block
         self.drop_block_cfg = drop_block_cfg
+        assert drop_block_cfg is None or isinstance(drop_block_cfg, dict)
 
         super().__init__(
             in_channels=[
@@ -98,24 +94,22 @@ class PPYOLOECSPPAN(BaseYOLONeck):
         """
         if idx == len(self.in_channels) - 1:
             # fpn_stage
-            channels = [
-                self.in_channels[idx]
-            ] + [self.out_channels[idx]] * self.num_stages_per_layer
+            channels = [self.in_channels[idx]
+                        ] + [self.out_channels[idx]] * self.num_csplayer
 
             layer = [
-                CSPResStage(
+                CSPResLayer(
                     in_channels=channels[i],
                     out_channels=channels[i + 1],
-                    num_block=self.num_blocks_per_stage,
+                    num_block=self.num_blocks_per_layer,
                     block_cfg=self.block_cfg,
                     norm_cfg=self.norm_cfg,
                     act_cfg=self.act_cfg,
-                    use_effective_se=False,
-                    use_spp=self.use_spp)
-                for i in range(self.num_stages_per_layer)
+                    effective_se_cfg=None,
+                    use_spp=self.use_spp) for i in range(self.num_csplayer)
             ]
 
-            if self.use_drop_block:
+            if self.drop_block_cfg:
                 layer.append(MODELS.build(self.drop_block_cfg))
             layer = nn.Sequential(*layer)
         else:
@@ -150,21 +144,21 @@ class PPYOLOECSPPAN(BaseYOLONeck):
         # fpn_stage
         in_channels = self.in_channels[idx - 1] + self.out_channels[idx] // 2
         out_channels = self.out_channels[idx - 1]
-        channels = [in_channels] + [out_channels] * self.num_stages_per_layer
+        channels = [in_channels] + [out_channels] * self.num_csplayer
 
         layer = [
-            CSPResStage(
+            CSPResLayer(
                 in_channels=channels[i],
                 out_channels=channels[i + 1],
-                num_block=self.num_blocks_per_stage,
+                num_block=self.num_blocks_per_layer,
                 block_cfg=self.block_cfg,
                 norm_cfg=self.norm_cfg,
                 act_cfg=self.act_cfg,
-                use_effective_se=False,
-                use_spp=False) for i in range(self.num_stages_per_layer)
+                effective_se_cfg=None,
+                use_spp=False) for i in range(self.num_csplayer)
         ]
 
-        if self.use_drop_block:
+        if self.drop_block_cfg:
             layer.append(MODELS.build(self.drop_block_cfg))
 
         return nn.Sequential(*layer)
@@ -200,21 +194,21 @@ class PPYOLOECSPPAN(BaseYOLONeck):
         # pan_stage
         in_channels = self.out_channels[idx + 1] + self.out_channels[idx]
         out_channels = self.out_channels[idx + 1]
-        channels = [in_channels] + [out_channels] * self.num_stages_per_layer
+        channels = [in_channels] + [out_channels] * self.num_csplayer
 
         layer = [
-            CSPResStage(
+            CSPResLayer(
                 in_channels=channels[i],
                 out_channels=channels[i + 1],
-                num_block=self.num_blocks_per_stage,
+                num_block=self.num_blocks_per_layer,
                 block_cfg=self.block_cfg,
                 norm_cfg=self.norm_cfg,
                 act_cfg=self.act_cfg,
-                use_effective_se=False,
-                use_spp=False) for i in range(self.num_stages_per_layer)
+                effective_se_cfg=None,
+                use_spp=False) for i in range(self.num_csplayer)
         ]
 
-        if self.use_drop_block:
+        if self.drop_block_cfg:
             layer.append(MODELS.build(self.drop_block_cfg))
 
         return nn.Sequential(*layer)
