+++ tests/test_models/test_backbone/test_csp_resnet.py
@@ -5,7 +5,7 @@ import pytest
 import torch
 from torch.nn.modules.batchnorm import _BatchNorm
 
-from mmyolo.models import CSPResNet
+from mmyolo.models import PPYOLOECSPResNet
 from mmyolo.utils import register_all_modules
 from .utils import check_norm_state, is_norm
 
@@ -17,16 +17,16 @@ class TestPPYOLOECSPResNet(TestCase):
     def test_init(self):
         # out_indices in range(len(arch_setting) + 1)
         with pytest.raises(AssertionError):
-            CSPResNet(out_indices=(6, ))
+            PPYOLOECSPResNet(out_indices=(6, ))
 
         with pytest.raises(ValueError):
             # frozen_stages must in range(-1, len(arch_setting) + 1)
-            CSPResNet(frozen_stages=6)
+            PPYOLOECSPResNet(frozen_stages=6)
 
     def test_forward(self):
-        # Test YOLOv6EfficientRep with first stage frozen
+        # Test PPYOLOECSPResNet with first stage frozen
         frozen_stages = 1
-        model = CSPResNet(frozen_stages=frozen_stages)
+        model = PPYOLOECSPResNet(frozen_stages=frozen_stages)
         model.init_weights()
         model.train()
 
@@ -41,14 +41,14 @@ class TestPPYOLOECSPResNet(TestCase):
             for param in layer.parameters():
                 assert param.requires_grad is False
 
-        # Test YOLOv6EfficientRep with norm_eval=True
-        model = CSPResNet(norm_eval=True)
+        # Test PPYOLOECSPResNet with norm_eval=True
+        model = PPYOLOECSPResNet(norm_eval=True)
         model.train()
 
         assert check_norm_state(model.modules(), False)
 
-        # Test YOLOv6EfficientRep-P5 forward with widen_factor=0.25
-        model = CSPResNet(
+        # Test PPYOLOECSPResNet-P5 forward with widen_factor=0.25
+        model = PPYOLOECSPResNet(
             arch='P5', widen_factor=0.25, out_indices=range(0, 5))
         model.train()
 
@@ -61,8 +61,8 @@ class TestPPYOLOECSPResNet(TestCase):
         assert feat[3].shape == torch.Size((1, 128, 4, 4))
         assert feat[4].shape == torch.Size((1, 256, 2, 2))
 
-        # Test YOLOv6EfficientRep forward with dict(type='ReLU')
-        model = CSPResNet(
+        # Test PPYOLOECSPResNet forward with dict(type='ReLU')
+        model = PPYOLOECSPResNet(
             widen_factor=0.125,
             act_cfg=dict(type='ReLU'),
             out_indices=range(0, 5))
@@ -77,8 +77,8 @@ class TestPPYOLOECSPResNet(TestCase):
         assert feat[3].shape == torch.Size((1, 64, 4, 4))
         assert feat[4].shape == torch.Size((1, 128, 2, 2))
 
-        # Test YOLOv6EfficientRep with BatchNorm forward
-        model = CSPResNet(widen_factor=0.125, out_indices=range(0, 5))
+        # Test PPYOLOECSPResNet with BatchNorm forward
+        model = PPYOLOECSPResNet(widen_factor=0.125, out_indices=range(0, 5))
         for m in model.modules():
             if is_norm(m):
                 assert isinstance(m, _BatchNorm)
@@ -93,8 +93,8 @@ class TestPPYOLOECSPResNet(TestCase):
         assert feat[3].shape == torch.Size((1, 64, 4, 4))
         assert feat[4].shape == torch.Size((1, 128, 2, 2))
 
-        # Test YOLOv6EfficientRep with BatchNorm forward
-        model = CSPResNet(plugins=[
+        # Test PPYOLOECSPResNet with BatchNorm forward
+        model = PPYOLOECSPResNet(plugins=[
             dict(
                 cfg=dict(type='mmdet.DropBlock', drop_prob=0.1, block_size=3),
                 stages=(False, False, True, True)),
