+++ tests/test_models/test_necks/test_ppyoloe_csppan.py
@@ -3,13 +3,13 @@ from unittest import TestCase
 
 import torch
 
-from mmyolo.models import PPYOLOECSPPAN
+from mmyolo.models import PPYOLOECSPPAFPN
 from mmyolo.utils import register_all_modules
 
 register_all_modules()
 
 
-class TestPPYOLOECSPPAN(TestCase):
+class TestPPYOLOECSPPAFPN(TestCase):
 
     def test_forward(self):
         s = 64
@@ -20,7 +20,7 @@ class TestPPYOLOECSPPAN(TestCase):
             torch.rand(1, in_channels[i], feat_sizes[i], feat_sizes[i])
             for i in range(len(in_channels))
         ]
-        neck = PPYOLOECSPPAN(
+        neck = PPYOLOECSPPAFPN(
             in_channels=in_channels, out_channels=out_channels)
         outs = neck(feats)
         assert len(outs) == len(feats)
@@ -37,10 +37,14 @@ class TestPPYOLOECSPPAN(TestCase):
             torch.rand(1, in_channels[i], feat_sizes[i], feat_sizes[i])
             for i in range(len(in_channels))
         ]
-        neck = PPYOLOECSPPAN(
+        neck = PPYOLOECSPPAFPN(
             in_channels=in_channels,
             out_channels=out_channels,
-            use_drop_block=True)
+            drop_block_cfg=dict(
+                type='mmdet.DropBlock',
+                drop_prob=0.1,
+                block_size=3,
+                warm_iters=0))
         neck.train()
         outs = neck(feats)
         assert len(outs) == len(feats)
